# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest Api documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://rest-api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Signup",
                "operationId": "signup",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignupResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/signup/activate/{id}": {
            "get": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Activate user",
                "operationId": "activateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignupResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LogoutResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
        "/api/auth/user": {
            "get": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Get current user info",
                "operationId": "currentUser",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrentUserResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": [
                            "*"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "LoginRequest": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "admin@admin.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the business",
                        "type": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
...
            "BadRequestResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "BadRequestResource",
                        "description": "Bad request resource",
                        "properties": {
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Bad request"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "BadRequestResource"
                        }
                    }
                ]
            },
            "ErrorResource": {
                "title": "ErrorResource",
                "description": "Error resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status wrapper",
                        "example": "error"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorResource"
                }
            },
            "SuccessResource": {
                "title": "SuccessResource",
                "description": "Success resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status wrapper",
                        "example": "success"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message wrapper",
                        "example": "The operation was successful."
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SuccessResource"
                }
            },
...
            "CurrentUserResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResource"
                    },
                    {
                        "title": "CurrentUserResource",
                        "description": "Current user resource",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "properties": {
                                    "user": {
                                        "title": "User",
                                        "description": "User wrapper",
                                        "properties": {
                                            "id": {
                                                "type": "int64",
                                                "example": 2
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "User"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "test@test.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "CurrentUserResource"
                        }
                    }
                ]
            },
...
            "ForbiddenResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "ForbiddenResource",
                        "description": "Forbidden resource",
                        "properties": {
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Forbidden"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "ForbiddenResource"
                        }
                    }
                ]
            },
            "LoginResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResource"
                    },
                    {
                        "title": "LoginResource",
                        "description": "Login resource",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "properties": {
                                    "loginData": {
                                        "title": "Login data",
                                        "description": "Login data wrapper",
                                        "properties": {
                                            "access_token": {
                                                "type": "string",
                                                "example": "IxIiwianRpIjoiN2I2NTYxNWM0ZWZmZGFiYWU..."
                                            },
                                            "token_type": {
                                                "type": "string",
                                                "example": "Bearer"
                                            },
                                            "expires_at": {
                                                "type": "string",
                                                "example": "2021-08-03 18:30:35"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "LoginResource"
                        }
                    }
                ]
            },
            "LogoutResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResource"
                    },
                    {
                        "title": "LogoutResource",
                        "description": "Logout resource",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "LogoutResource"
                        }
                    }
                ]
            },
            "NotFoundResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "NotFoundResource",
                        "description": "Not found resource",
                        "properties": {
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Not found"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "NotFoundResource"
                        }
                    }
                ]
            },
            "SignupResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/SuccessResource"
                    },
                    {
                        "title": "SignupResource",
                        "description": "Signup resource",
                        "properties": {
                            "data": {
                                "title": "Data",
                                "description": "Data wrapper",
                                "properties": {
                                    "user": {
                                        "title": "User data",
                                        "description": "User data wrapper",
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "User"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "test@test.com"
                                            },
                                            "id": {
                                                "type": "int64",
                                                "example": 2
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "SignupResource"
                        }
                    }
                ]
            },
            "UnauthenticatedResource": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ErrorResource"
                    },
                    {
                        "title": "UnauthenticatedResource",
                        "description": "Unauthenticated resource",
                        "properties": {
                            "status": {
                                "title": "Status",
                                "description": "Status wrapper",
                                "example": "failed"
                            },
                            "message": {
                                "title": "Message",
                                "description": "Message wrapper",
                                "example": "Unauthenticated"
                            }
                        },
                        "type": "object",
                        "xml": {
                            "name": "UnauthenticatedResource"
                        }
                    }
                ]
            },
...
            "SignupRequest": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the user",
                        "type": "string",
                        "example": "User"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password of the business",
                        "type": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "Password confirmation",
                        "type": "password",
                        "example": "password"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://rest-api/oauth/authorize",
                        "tokenUrl": "http://rest-api/oauth/token",
                        "refreshUrl": "http://rest-api/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}